pyproject.toml
[tool.poetry]
name = "github_actions"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2 = "^2.9.3"
Django = "^4.0.5"
django-rest-framework = "^0.1.0"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
flake8 = "^4.0.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "github_actions.settings"
addopts = "-p no:cacheprovider"